% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factors}
\alias{factors}
\title{Factor modeling: Inference for the number of factors}
\usage{
factors(Y, lag.k = 5, twostep = FALSE)
}
\arguments{
\item{Y}{A data matrix used for factor inference with \eqn{n} rows and \eqn{p} columns, where \eqn{n} is the sample size and \eqn{p} is the dimension of the time series.}

\item{lag.k}{A time lag which is a positive integer specified to calculate \eqn{ \hat{\mathbf{M}}}.\deqn{\mathbf{M}\ =\ \sum_{k=1}^{k_0}\mathbf{\Sigma}_y(k){\mathbf{\Sigma}_y(k)}^{'} }}

\item{twostep}{logical. If \code{FALSE} (the default), then standard procedures for factor number inference will be implemented. If \code{TRUE}, then a two step estimation procedure will be implemented 
for the estimation of weak factors.}
}
\value{
An object of class "factors" is a list containing the following components:

\item{factor_num}{Number of factors \eqn{r}}
\item{loading.mat}{The estimated \eqn{p \times r} factor loading matrix}
}
\description{
\code{factors()} deals with factor modeling for high-dimensional time series based on a dimension-reduction 
viewpoint. Under stationary settings, the inference is simple in the sense that both the number of factors \eqn{r} 
and the factor loadings \eqn{A} are estimated in terms of an eigenanalysis for a nonnegative definite matrix, and is 
therefore applicable when the dimension of time series is on the order of a few thousands.
}
\examples{
## Generate x_t 
p <- 400
n <- 400
r <- 3
X <- mat.or.vec(n, r)
A <- matrix(runif(p*r, -1, 1), ncol=r)
x1 <- arima.sim(model=list(ar=c(0.6)), n=n)
x2 <- arima.sim(model=list(ar=c(-0.5)), n=n)
x3 <- arima.sim(model=list(ar=c(0.3)), n=n)
eps <- matrix(rnorm(n*p), p, n)
X <- t(cbind(x1, x2, x3))
Y <- A \%*\% X + eps
Y <- t(Y)
fac <- factors(Y,lag.k=2)
r_hat <- fac$factor_num
loading_Mat <- fac$loading.mat
}
\references{
Clifford Lam. Qiwei Yao. \emph{Factor modelling for high-dimensional time series: inference for the number of factors}. The Annals of Statistics, 2012, 40, 694-726.
}
