% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FamaFrench}
\alias{FamaFrench}
\title{Fama-French 10*10 return series}
\format{
A data frame with 696 rows and 102 columns. The first column
represents the month, and the second column named
\code{MKT.RF} represents the monthly market returns. The rest of the columns
represent the return series for different sizes and BE-ratios.
}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
data(FamaFrench)
}
\description{
The portfolios are constructed by the intersections of 10 levels
of size, denoted by \eqn{{\rm S}_{1}, \ldots, {\rm S}_{10}}, and 10 levels of the book
equity to market equity ratio (BE), denoted by \eqn{{\rm BE}_1, \ldots,{\rm BE}_{10}}.
The dataset consists of monthly returns from January 1964 to
December 2021, which contains 69600 observations for 696 total months.
}
\keyword{data}
