% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedimensionUR.R
\name{onedimensionUR}
\alias{onedimensionUR}
\title{onedimensionUR}
\usage{
onedimensionUR(Z, K0_vec, con_vec = con_vec)
}
\arguments{
\item{Z}{a time series}

\item{K0_vec}{lag K0 vector, if missing we use K0_vec=c(1,2,3,4,5)}

\item{con_vec}{constant vector for ck}
}
\value{
A dataframe "1" means we reject the null hypothesis.
}
\description{
You can use this r code function to test unit root series
}
\examples{
N=100
Z=arima.sim(list(ar=c(0.9)), n = 2*N, sd=sqrt(1))
con_vec=c(0,0.45,0.55,0.65)
onedimensionUR(Z, con_vec=con_vec)
}
