% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFDR.R
\name{permutationFDR}
\alias{permutationFDR}
\title{Permutation Using the FDR Method}
\usage{
permutationFDR(X, isvol = FALSE, Beta = NULL, m = NULL)
}
\arguments{
\item{X}{A data matrix used to find the grouping mechanism with \eqn{n} rows and \eqn{p} columns, where \eqn{n} is the sample size and \eqn{p} is the dimension of the time series.}

\item{isvol}{Logical. If \code{FALSE} (the default), then prewhiten each series by fitting a univariate AR model with
the order between 0 and 5 determined by AIC. If \code{TRUE}, then prewhiten each volatility process using GARCH(1,1) model.}

\item{Beta}{A positive constant}

\item{m}{A positive constant used to calculate the maximum cross correlation over the lags between \eqn{-m} and \eqn{m}. If \eqn{m} is not specified, the default constant \eqn{10*log10(n/p)}
will be used.}
}
\value{
An object of class "permutationFDR" is a list containing the following components:

\item{NoGroups}{Number of groups with at least two components series}
\item{Nos_of_Members}{Number of members in each of groups with at least two members}
\item{Groups}{Indices of components in each of groups with at least two members}
\item{maxcorr}{Maximum correlation (over lags) of \eqn{p(p-1)/2} pairs in descending order}
\item{corrRatio}{Ratios of successive values from maxcorr}
\item{NoConnectedPairs}{Number of connected pairs}
\item{Xpre}{The prewhitened data with \eqn{n-R} rows and \eqn{p} columns}
}
\description{
The permutation is determined by grouping the components of a multivariate series X into \eqn{q} groups, where \eqn{q} and the cardinal numbers of those groups are also unknown. This function use FDR Method
}
\details{
See Chang et al. (2018) for the permutation step and more information.
}
\note{
This is the second step for segmentation by grouping the transformed time series. The first step is to seek for a contemporaneous linear transformation of the original series, see \code{\link{segmentTS}}.
}
\references{
Chang, J., Guo, B. & Yao, Q. (2018). \emph{Principal component analysis for second-order stationary vector time series}. The Annals of Statistics, Vol. 46, pp. 2094–2124.
}
\seealso{
\code{\link{segmentTS}}, \code{\link{permutationMax}}
}
