% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mtscp}
\alias{predict.mtscp}
\title{Make predictions from a \code{"mtscp"} object.}
\usage{
\method{predict}{mtscp}(
  object,
  newdata = NULL,
  n.ahead = 10,
  control_ARIMA = list(),
  control_VAR = list(),
  ...
)
}
\arguments{
\item{object}{An object of class \code{"mtscp"} constructed by \code{\link{CP_MTS}}.}

\item{newdata}{A new data matrix to predict from (optional).}

\item{n.ahead}{An integer specifying the number of steps ahead to predict.}

\item{control_ARIMA}{A list of arguments passed to the function
\code{auto.arima()} of \pkg{forecast} [See the manual of \code{auto.arima()}].
See 'Details'. The default is \code{list(ic = "aic")}.}

\item{control_VAR}{A list of arguments passed to the function
\code{VAR()} of \pkg{vars} [See the manual of \code{VAR()}]. See 'Details'.
The default is \code{list(type = "const", lag.max = 6, ic = "AIC")}.}

\item{...}{Currently not used.}
}
\value{
\code{Y_pred} A list of length \code{n.ahead}, where each element is a 
\eqn{p \times q} matrix representing the predicted values for 
each time steps.
}
\description{
This function makes predictions from a \code{"mtscp"} object.
}
\details{
Forecasting for \eqn{{\bf y}_t} can be implemented in two steps:

\emph{Step 1}. Get the \eqn{h}-step ahead forecast of the \eqn{\hat{d} \times 1} 
time series \eqn{\hat{\bf x}_t} [See \code{\link{CP_MTS}}] by using a VAR model 
(if \eqn{\hat{d} > 1}) or an ARIMA model
(if \eqn{\hat{d} = 1}). The orders of VAR and ARIMA models are determined by
AIC by default. Otherwise, they can also be specified by users through the
arguments \code{control_VAR} and \code{control_ARIMA}, respectively.

\emph{Step 2}. The forecasted value for \eqn{{\bf y}_t} is obtained by
 \eqn{\hat{\bf y}_{n+h}= \hat{\bf A} {\rm diag}(\hat{\bf x}_{n+h}) \hat{\bf B}'} .
}
\seealso{
\code{\link{CP_MTS}}
}
