% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentTS.R
\name{segmentTS}
\alias{segmentTS}
\title{Segment Multivariate Time Series}
\usage{
segmentTS(Y, lag.k, isvol = FALSE, thresh = FALSE, tuning.vec = NULL, K = 5)
}
\arguments{
\item{Y}{A data matrix with \eqn{n} rows and \eqn{p} columns, where \eqn{n} is the sample size and \eqn{p} is the dimension of the time series.}

\item{lag.k}{A positive integer specified to calculate Wy. See (2.5) in Chang et al. (2018).}

\item{isvol}{Logical. If \code{FALSE} (the default), it means that the process is stationary, otherwise it is Volatility process.}

\item{thresh}{Logical. If \code{FALSE} (the default), no thresholding will be applied. If \code{TRUE}, a
thresholding method will be applied first to estimate Wy, see (3.4) and (3.5) in Chang et al. (2018).}

\item{tuning.vec}{The value of thresholding parameter \eqn{\lambda}. The thresholding level is specified by
\deqn{ u = \lambda {(log p/n)^(1/2)}.}
  Default value is 2. If \code{tuning.vec} is a vector, then a cross validation method proposed in Cai and Liu (2011) will be used
  to choose the best tuning parameter.}

\item{K}{The number of folders used in the cross validation, the default is 5. It is required when \code{thresh} is \code{TRUE}.}
}
\value{
An object of class "segmentTS" is a list containing the following components:
\item{B}{the \eqn{p} by \eqn{p} transformation matrix such that \eqn{x_t=By_t}}
\item{X}{the transformed series with \eqn{n} rows and \eqn{p} columns}
}
\description{
Calculate linear transformation of the \eqn{p}-variate time series \eqn{y_t} such that the transformed series \eqn{x_t=By_t} is segmented into several
lower-dimensional subseries, and those subseries are uncorrelated with
each other both contemporaneously and serially.
}
\details{
When \eqn{p} is small, thresholding is not required. However, when \eqn{p} is large, it is necessary to use the thresholding method, see more information in Chang et al. (2014).
}
\note{
This is the first step to transform the time series. The second step is grouping the transformed time series, see \code{\link{permutationMax}}, \code{\link{permutationFDR}}.
}
\references{
Chang, J., Guo, B. & Yao, Q. (2018). \emph{Principal component analysis for second-order stationary vector time series}. The Annals of Statistics, Vol. 46, pp. 2094–2124.

            Cai, T. and Liu, W. (2011). \emph{Adaptive thresholding for sparse covariance matrix estimation}.  Journal of the American Statistical Association 106: 672-684.
}
\seealso{
\code{\link{permutationMax}}, \code{\link{permutationFDR}}.
}
