% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools_SpecTest.R
\name{SpecTest}
\alias{SpecTest}
\title{Statistical inference for high-dimensional spectral density matrix}
\usage{
SpecTest(B.monte, flag_c, J.set, n, p, x, cross.indices)
}
\arguments{
\item{B.monte}{Bootstrap times for generating multivariate normal distributed 
random vectors in calculating the critical value.}

\item{flag_c}{Bandwidth \eqn{c} of the flat-top kernel for estimating 
\eqn{f_{i,j}(\omega)}, where \eqn{c\in(0,1]}.}

\item{J.set}{Set \eqn{\mathcal{J}} for frequencies, a vector, used to calculate the test statistic.}

\item{n}{Sample size.}

\item{p}{Dimension of \eqn{{\bf x}_t}.}

\item{x}{\eqn{{\bf x} = \{{\bf x_1}, \dots , {\bf x}_n \}}, a \eqn{p\times
n} sample matrix, where \eqn{n} is the sample size and \eqn{p} is the 
dimension of \eqn{{\bf x}_t}.}

\item{cross.indices}{Set \eqn{\mathcal{I}} for \eqn{(i,j)}, a matrix with 2 columns,
used to calculate the test statistic.}
}
\value{
An object of class "SpecTest" is a list containing the following
  components:

  \item{Stat}{Numerical value which represents the value of test statistic.}
  \item{pval}{Numerical value which represents the p-value of the test.}
  \item{cri95}{Numerical value which represents the critical value of the test
  at the significance level 0.05.}
}
\description{
\code{SpecTest()} implements a new global test proposed in
 Chang, Jiang, McElroy and Shao (2023) for the following hypothesis testing problem: 
\deqn{H_0:f_{i,j}(\omega)=0 \mathrm{\ for\ any\ }(i,j)\in\mathcal{I}\mathrm{\ and\ }
\omega \in \mathcal{J}\mathrm{\ \ versus\ \ H_1:H_0\ is\ not\ true.} }
}
\examples{
n <- 200
p <- 10
flag_c <- 0.8
B.monte <- 1000
burn <- 1000
z.sim <- matrix(rnorm((n+burn)*p),p,n+burn)
phi.mat <- 0.4*diag(p)
x.sim <- phi.mat \%*\% z.sim[,(burn+1):(burn+n)]
x <- x.sim - rowMeans(x.sim)
cross.indices <- matrix(c(1,2), ncol=2)
J.set <- 2*pi*seq(0,3)/4 - pi
res <- SpecTest(B.monte, flag_c, J.set, n, p, x, cross.indices)
Stat <- res$Stat
Pvalue <- res$pval
CriVal <- res$cri95
}
\references{
Chang, J., Jiang, Q., McElroy, T. & Shao, X. (2023). 
\emph{Statistical inference for high-dimensional spectral density matrix}.
}
