% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urtest.R
\name{UR_test}
\alias{UR_test}
\title{Testing for unit roots based on sample autocovariances}
\usage{
UR_test(Y, lagk.vec = NULL, con_vec = NULL, alpha = 0.05)
}
\arguments{
\item{Y}{A vector \eqn{Y = (y_1, \dots , y_n )}, where \eqn{n} is the number
of the observations.}

\item{lagk.vec}{A vector of length \eqn{d} with each element representing a
specified time lag \eqn{K_0} used to calculate the test statistic
[See Section 2.1 of Chang, Cheng and Yao (2022)].
The procedure outputs all test results corresponding to the specified \eqn{d} lags.
The default is \code{c(0, 1, 2, 3, 4)}.}

\item{con_vec}{A vector of length \eqn{m} with each element representing a
specified constant \eqn{c_\kappa} [See (5) of Chang, Cheng and Yao (2022)]. 
The procedure outputs all test results corresponding to \eqn{m} values of \eqn{c_\kappa}.
The default is \code{c(0.55)}.}

\item{alpha}{The significance level of the test. Default is 0.05.}
}
\value{
An object of class \code{"urtest"}, which contains the following
  components:

  \item{statistic}{A vector of length \eqn{d} with each element representing
  the value of a test statistic corresponding to one of the \eqn{d} time lags.}
  \item{reject}{An \eqn{m \times d} data matrix containing test results for
  various parameter (\eqn{c_\kappa,K_0}) combinations. Each row
   corresponds to a specific value of \eqn{c_\kappa} and 
   each column represents a specific time lag \eqn{K_0},  A value of
   \code{'1'} indicates rejection of the null hypothesis,
   while \code{'0'} indicates non-rejection.}
  \item{lag.vec}{The time lag used in function.}
}
\description{
The test implements the test proposed in Chang, Cheng and Yao (2022)
for the following hypothesis testing problems:
\deqn{H_0:Y_t \sim I(0)\ \ \mathrm{versus}\ \ H_1:Y_t \sim
I(d)\ \mathrm{for\ some\ integer\ }d \geq 2\,,} where \eqn{\{Y_t\}_{t=1}^n} is
a univariate time series.
}
\examples{
# Example 1
## generate yt
N <- 100
Y <-arima.sim(list(ar = c(0.9)), n = 2*N, sd = sqrt(1))
con_vec <- c(0.45, 0.55, 0.65)
lagk.vec <- c(0, 1, 2)

UR_test(Y, lagk.vec = lagk.vec, con_vec = con_vec, alpha = 0.05)
UR_test(Y, alpha = 0.05)
}
\references{
Chang, J., Cheng, G., & Yao, Q. (2022). Testing for Unit
  Roots Based on Sample Autocovariances. \emph{Biometrika}, \strong{109},
  543--550. \doi{doi:10.1093/biomet/asab034}
}
