% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPindices}
\alias{IPindices}
\title{US Industrial Production indices}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{var1}{Description of var1}
  \item{var2}{Description of var2}
  \item{var3}{Description of var3}
}
}
\usage{
data(IPindices)
}
\description{
this is an example dataset for demonstration purposes.
}
\keyword{data}
