% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_functions_unified.R
\name{CP_Inference}
\alias{CP_Inference}
\title{Inference for the estimates obtained by Double Projection Iterations (DPI) method in the tensor CP-factor model}
\usage{
CP_Inference(h, i, j, Y, res.CP.DPI)
}
\arguments{
\item{h}{A numeric vector of length \eqn{d_j}. This specifies the linear
functional \eqn{\mathbf{h}^\top (\hat{\mathbf{a}}_{i,j} - \hat{\mathbf{\vartheta}}_{i.j})} of interest.}

\item{i}{A positive integer. The factor index.}

\item{j}{A positive integer. The tensor mode index.}

\item{Y}{An array containing the observed tensor time series. Its dimension is
typically \eqn{n \times d_1 \times \cdots \times d_m}.}

\item{res.CP.DPI}{An output object from \code{CP_MTS} with \code{method = "CP.DPI"}.}
}
\value{
A list with the following components:
\describe{
  \item{\code{aij.h.de}}{ \eqn{\mathbf{h}^\top (\hat{\mathbf{a}}_{i,j} - \hat{\mathbf{\vartheta}}_{i.j})}.}
  \item{\code{se.h.ij}}{The estimated standard error of \code{aij.h.de}.}
  \item{\code{aij.iter}}{The original DPI estimator of the loading vector \eqn{\mathbf{a}_{i,j}}.}
  \item{\code{vartheta.ij}}{The estimated bias-correction term \eqn{\hat{\mathbf{\vartheta}}_{i.j}} used in debiasing.}
}
}
\description{
This function performs inference for a linear functional of the debiased
DPI estimator of the factor loading vector in the tensor CP-factor model.
Given a direction vector \code{h}, a factor index \code{i}, a mode index \code{j},
the observed tensor series \code{Y}, and the output object \code{res.CP.DPI} of the DPI method,
the function returns the debiased estimate of \eqn{\mathbf{h}^\top (\hat{\mathbf{a}}_{i,j} - \hat{\mathbf{\vartheta}}_{i.j})},
its estimated standard error, the corresponding DPI estimator \eqn{\hat{\mathbf{a}}_{i,j}}, and the
estimated bias-correction term \eqn{\hat{\mathbf{\vartheta}}_{i.j})}.
}
\details{
The reported standard error is
\deqn{
\sqrt{\widehat{\mathrm{Var}} \big\{\mathbf{h}^\top (\hat{\mathbf{a}}_{i,j} - \hat{\mathbf{\vartheta}}_{i.j})\big\}/n} \,,
}
where \eqn{n} is the sample size.
}
\examples{
\dontrun{
out <- CP_Inference(
  h = h,
  i = 1,
  j = 2,
  Y = Y,
  res.CP.DPI = res.CP.DPI
)

out$aij.h.de
out$se.h.ij
out$aij.iter
out$vartheta.ij
}

}
