% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools_martingale.R
\name{MartG_test}
\alias{MartG_test}
\title{Testing for martingale difference hypothesis in high dimension}
\usage{
MartG_test(
  X,
  lag.k = 2,
  B = 1000,
  type = c("Linear", "Quad", "In"),
  alpha = 0.05,
  kernel.type = c("QS", "BR", "BT")
)
}
\arguments{
\item{X}{An \eqn{n\times p} data matrix used in testing whether the series are martingale difference process.}

\item{lag.k}{A time lag which is a positive integer. Default is \code{lag.k} \eqn{=2}.}

\item{B}{Bootstrap times for generating vector from a multivariate normal distribution. Default is \code{B} \eqn{=2000}}

\item{type}{String, a map is chosen by the \proglang{R} users, such as the default option is \code{'Linear'} means linear identity map. Also including \code{Quad} and \code{ln}}

\item{alpha}{The significance level used for testing. Default is 0.05.}

\item{kernel.type}{String, an option for choosing an optimal symmetric kernel type, for example, \code{QS}, \code{Par}, \code{Bart}. Default is \code{kernel.type ='QS'}.}
}
\value{
An object of class "MartG_test" is a list containing the following components:

\item{reject}{\code{Logic} value which represents whether the process are martingale difference process. }
\item{p.value}{Numerical type which represents the \eqn{p}-value obtained in the bootstrap sampling process}
}
\description{
\code{MartG_test()} implements a new test for the martingale difference hypothesis for high-dimensional time series. 
The test is built on the sup-norm of a matrix-valued sample nonlinear dependencemetric at a finite and possibly growing number of lags. 
To the best of knowledge, this is the first valid test for the martingale difference hypothesis that allows for large dimension 
and captures nonlinear serial dependence. Besides, the test is robust to conditional moments of unknown forms and panel dependence of 
unknown magnitude.
}
\examples{
n <- 200
p <- 150
X <- matrix(rnorm(n*p),n,p)
res <- MartG_test(X)
Pvalue <- res$p.value
rej <- res$reject
}
\references{
Chang, J., Jiang, Q. & Shao, X. (2021). \emph{Testing the martingale difference hypothesis in high dimension}.
}
