% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA4_TS.R
\name{PCA4_TS}
\alias{PCA4_TS}
\title{Principal component analysis for time serise}
\usage{
PCA4_TS(
  Y,
  lag.k = 5,
  thresh = FALSE,
  tuning.vec = NULL,
  K = 5,
  isvol = FALSE,
  permutation = c("max", "fdr"),
  m = NULL,
  Beta = NULL,
  just4pre = FALSE
)
}
\arguments{
\item{Y}{A data matrix with \eqn{n} rows and \eqn{p} columns, where \eqn{n} is the sample size and \eqn{p} is the dimension of the time series.}

\item{lag.k}{A positive integer specified to calculate Wy. See (2.5) in Chang et al. (2014).}

\item{thresh}{Logical. If \code{FALSE} (the default), no thresholding will be applied. If \code{TRUE}, a
thresholding method will be applied first to estimate Wy, see (3.4) and (3.5) in Chang et al. (2018).}

\item{tuning.vec}{The value of thresholding parameter \eqn{\lambda}. The thresholding level is specified by
\deqn{ u = \lambda {(log p/n)}^{1/2}.}
  Default value is 2. If \code{tuning.vec} is a vector, then a cross validation method proposed in Cai and Liu (2011) will be used
  to choose the best tuning parameter.}

\item{K}{The number of folders used in the cross validation, the default is 5. It is required when \code{thresh} is \code{TRUE}.}

\item{isvol}{Logical. If \code{FALSE} (the default), then prewhiten each series by fitting a univariate AR model with
the order between 0 and 5 determined by AIC. If \code{TRUE}, then prewhiten each volatility process using GARCH(1,1) model.}

\item{permutation}{The method of permutation procdure,the option is \code{'max'} or \code{'fdr'},default option is \code{'max'}}

\item{m}{A positive constant used to calculate the maximum cross correlation over the lags between \eqn{-m} and \eqn{m}. If \eqn{m} is not specified, the default constant \eqn{10*log10(n/p)}
will be used.}

\item{Beta}{The error rate in FDR.}

\item{just4pre}{Logical. If \code{TRUE}, then only segment procdure will be used, If \code{FALSE} (the default), the nomal procdure will be used.}
}
\value{
The first step of segment containing the following components:
\item{B}{The \eqn{p*p} transformation matrix such that \eqn{x_t = By_t}}
\item{X}{The transformed series with \eqn{n} rows and \eqn{p} columns}

Output of the second step is a list containing the following components:
\item{NoGroups}{Number of groups with at least two components series}
\item{Nos_of_Members}{Number of members in each of groups with at least two members}
\item{Groups}{Indices of components in each of groups with at least two members}
\item{maxcorr}{Maximum correlation (over lags) of \eqn{p(p-1)/2} pairs in descending order if \code{permutation} is \code{max}}
\item{corrRatio}{Ratios of successive values from maxcorr if \code{permutation} is \code{max}}
\item{Pvalues}{Pvalue for multiple test H_0 for each of \eqn{p(p-1)/2} pairs in ascending order if \code{permutation} is \code{fdr}}
\item{NoConnectedPairs}{Number of connected pairs}
\item{Xpre}{The prewhitened data with \eqn{n-R} rows and \eqn{p} columns}
}
\description{
\code{PCA4_TS()} seek for a contemporaneous linear transformation for
a multivariate time series such that the transformed series is segmented
into several lower-dimensional subseries, and those subseries are
uncorrelated with each other both contemporaneously and serially, then permutation step can
grouping the components of a multivariate series X into \eqn{q} groups.
}
\details{
Before segment step, the time series need to be normalized then the variance of \eqn{Y}_{t} will be \eqn{I}_{p}. 
 When \eqn{p} is large (\eqn{>n}),it is necessary to use the package \code{clime}. When \eqn{p} is large (\eqn{p=O(n^{1/2}})), it is necessary to use the thresholding method, see more information in Chang et al. (2018).
}
\note{
The first step is transform the time series. It calculate linear transformation of the \eqn{p}-variate time series (or volatility processes) \eqn{y_t} such that the transformed series \eqn{x_t=By_t} is segmented into several
lower-dimensional subseries, and those subseries are uncorrelated with each other both contemporaneously and serially.
The second step is grouping the transformed time series, The permutation is determined by grouping the components of a multivariate series X into \eqn{q} groups, where \eqn{q} and the cardinal numbers of those groups are also unknown. See also \code{\link{permutationFDR}} and \code{\link{permutationMax}}
}
\references{
Chang, J., Guo, B. & Yao, Q. (2018). \emph{Principal component analysis for second-order stationary vector time series}. The Annals of Statistics, Vol. 46, pp. 2094–2124.

            Cai, T. and Liu, W. (2011). \emph{Adaptive thresholding for sparse covariance matrix estimation}.  Journal of the American Statistical Association 106: 672-684.

            Cai, T.T., Liu, W., and Luo, X. (2011). \emph{A constrained `1 minimization approach for sparse precision matrix estimation}. Journal of the American Statistical Association 106(494): 594-607.
}
\seealso{
\code{\link{segmentTS}}, \code{\link{permutationMax}}, \code{\link{permutationFDR}}.
}
