% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urtest.R
\name{ur.test}
\alias{ur.test}
\title{ur.test}
\usage{
ur.test(Z, K0_vec = K0_vec, con_vec = con_vec, alpha = alpha)
}
\arguments{
\item{Z}{input one dimensional time series}

\item{K0_vec}{lag K0 vector, if missing, the default value we choose K0_vec=c(0,1,2,3,4)}

\item{con_vec}{constant vector for ck, if missing, the default value we use 0.55}

\item{alpha}{significance level}
}
\value{
A dataframe, "1" means we reject the null hypothesis and "0" means we do not reject the null hypothesis
}
\description{
You can use this r code function urt to test for the existence of unit roots
}
\examples{
N=100
Z=arima.sim(list(ar=c(0.9)), n = 2*N, sd=sqrt(1))
con_vec=c(0.45,0.55,0.65)
K0_vec=c(0,1,2)
ur.test(Z,K0_vec=K0_vec, con_vec=con_vec,alpha=0.05)
ur.test(Z,alpha=0.05)
}
