% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPindices}
\alias{IPindices}
\title{U.S. Industrial Production indices}
\format{
A data frame with 924 rows and 8 variables:
\describe{
  \item{DATE}{The observation date}
  \item{INDPRO}{The total index}
  \item{IPB54000S}{Nonindustrial supplies}
  \item{IPFINAL}{Final products}
  \item{IPMANSICS}{Manufacturing}
  \item{IPMAT}{Materials}
  \item{IPMINE}{Mining}
  \item{IPUTIL}{Utilities}
}
}
\source{
\url{https://fred.stlouisfed.org/release/tables?rid=13&eid=49670}
}
\usage{
data(IPindices)
}
\description{
The dataset consists of 7 monthly U.S. Industrial
Production indices, namely \emph{the total index}, \emph{nonindustrial supplies},
\emph{final products}, \emph{manufacturing}, \emph{materials}, \emph{mining},
and \emph{utilities}, from January 1947 to December 2023 published by the U.S.
Federal Reserve.
}
\keyword{data}
